useGlobalEndCallKey hook
