+++ src/hooks/index.js
@@ -7,6 +7,7 @@ export * from './useArticleSummary'
 export * from './useConfirmDialog'
 export * from './useErrorLogging'
 export * from './useExperimentConfig'
+export * from './useGlobalEndCallKey'
 export * from './useHistoryState'
 export * from './useI18n'
 export * from './useNavigation'
